context: allow to specify a context during the put (out.sh)
This commit allow you to override context for each put.
If you don't specify context param during the put, the default one
from the resource configuration will be used.

The goal is to be able to declare one concourse resource to report status for
multiple jobs

For example :

* Job unittest : context unittest
* Job functionnal-tests: context functionnal-tests

The result of that in github will display a status for each jobs in your CI
as https://raw.githubusercontent.com/quinkennedy/travis-github-lint-status/master/example.png